requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
absl-py
astunparse
cachetools
certifi
chardet
docker
docker-compose
flask-cors
Gunicorn
flatbuffers
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
Keras-Preprocessing
Markdown
numpy
oauthlib
opt-einsum
Pillow
protobuf==3.20.*
pyasn1
pyasn1-modules
requests
requests-oauthlib
rsa
six
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
typing-extensions
urllib3
wrapt